Rename InstructionParser to OperationParser
